Refactoring minion task generator to reduce duplicate logic
